internal/interval: support different interval kinds
It might sometimes be useful to use fully open or fully closed
intervals. This supports that. For simplicity, we do this by converting
between other interval types and the existing half-open one. We use that
as the canonical because

1. It is the most natural interval for languages like Go and.
2. We already tested that.

In places, we have to use type-assertions to do these conversions,
because of limitations in Go generics.