internal/math: clean up ChineseRemainder API
Re-ordering the arguments so remainder and moduli are next to each other
makes their relationship a bit clearer. Also return the LCM, which is
almost always needed anyways.

Now, the Chinese remainder for a sequence can be calculated as

R, M := 0, 1
for r, m := range seq {
    R, M, _ = math.ChineseRemainder(R, M, r, m)
}

which is pretty clean.

Also add a version for big integers.